Move context.ts to helper files
