+++ playwright/test/search.test.ts
@@ -1,5 +1,5 @@
 import { Page, test, expect } from '@playwright/test';
-import { gotoWithoutCache, isMobile } from './contexts';
+import { gotoWithoutCache, isMobile } from './helpers/contexts';
 
 import { elementIsVisible, fillInputAction } from './actions/common';
 import { expectItemsIsVisible, expectItemIsVisible } from './asserts/common';
